Rename XMLPreferenceConstants -> XMLPreferenceServerConstants
